+++ GUI-dev/src/app/pages/approval-details/approval-details.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+
+
+import { ThemeModule } from '../../@theme/theme.module';
+
+
+@NgModule({
+  imports: [
+    ThemeModule,
+  ],
+  declarations: [
+    // ApprovalDetailsComponent,
+  ],
+})
+export class ApprovalDetailsModule { }