+++ GUI-dev/src/app/pages/pages-menu.ts
@@ -7,6 +7,11 @@ export const MENU_ITEMS: NbMenuItem[] = [
     link: '/dashboard',
   },
 
+  {
+    title: 'Approval Page',
+    icon: 'nb-checkmark',
+    link: '/approval',
+  },
   /*{
     title: 'FEATURES',
     group: true,