+++ GUI-dev/src/app/pages/pages-routing.module.ts
@@ -4,7 +4,8 @@ import { NgModule } from '@angular/core';
 import { PagesComponent } from './pages.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { HomeComponent } from './home/home.component';
-import { AuthGuard } from '../auth-guard.service';
+import {AuthGuard} from '../auth-guard.service';
+import {ApprovalComponent} from './approval/approval.component';
 import {ActiveProcessesComponent} from './dashboard/components/activeProcesses/activeProcesses.component';
 // import {EventLoggerComponent} from './admin/components/eventLogger/eventLogger.component';
 
@@ -30,6 +31,11 @@ const routes: Routes = [{
     {
       path: 'home',
       component: HomeComponent,
+    },
+    {
+      path: 'approval',
+      component: ApprovalComponent,
+      canActivate: [AuthGuard],
     },
     {
       path: '',